refactor: split ast into multiple files
